requirements/common.txt
django==1.8.16
pycryptodome==3.4.7